Fixed any inconsitencies of styles between browsers.
I added plus and minus icons and positioning them relative to a .slider-border div, which I also styled.

I wanted the rgb amount to follow the slider thumb as it moved, but could only get it to apply to Safari.

*Content added to the :before and :after psuedo-classes are only ment to be applied to elements which take content. Seeing as the slider-thumb isn't content editable, it doesn't accept the ::before and ::after. That it works in Safari is a bug.
Because of this I have the rgb amount box stationary centered in the slider.
I will change that whe I figure out how to do it at another time.